video stabilizer app is working
