+++ k3d/video_stabilizer_pod.yaml
@@ -1,0 +1,19 @@
+apiVersion: v1
+kind: Namespace
+metadata:
+  name: video-stabilizer-grpc
+
+---
+apiVersion: v1
+kind: Pod
+metadata:
+  name: video-stabilizer-pod
+  namespace: video-stabilizer-grpc
+  labels:
+    app: video-stabilizer
+spec:
+  containers:
+    - name: video-stabilizer
+      image: video-stabilizer:latest
+      imagePullPolicy: Never
+  restartPolicy: Never
