+++ video_stabilizer_clients/stabilize_client.py
@@ -6,7 +6,8 @@ class StabilizeClient(object):
     # Client for gRPC functionality
 
     def __init__(self):
-        self.host = 'sts-service.stabilize-server-grpc.svc.cluster.local'
+        # self.host = 'sts-service.stabilize-server-grpc.svc.cluster.local'
+        self.host = 'localhost'
         self.server_port = 50051
 
         # instantiate a channel
